core: fix RSE protocol sorting; Fix #2363

+++ lib/rucio/db/sqla/migrate_repo/versions/35ef10d1e11b_change_index_on_table_requests.py
@@ -1,39 +1,47 @@
-"""
- Copyright European Organization for Nuclear Research (CERN)
+# Copyright 2013-2019 CERN for the benefit of the ATLAS collaboration.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Authors:
+# - Vincent Garonne <vincent.garonne@cern.ch>, 2014-2017
+# - Mario Lassnig <mario.lassnig@cern.ch>, 2019
 
- Licensed under the Apache License, Version 2.0 (the "License");
- You may not use this file except in compliance with the License.
- You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+''' change index on table requests '''
 
- Authors:
- - Vincent Garonne, <vincent.garonne@cern.ch>, 2014-2017
-
-change index on table requests
-
-Revision ID: 35ef10d1e11b
-Revises: 3152492b110b
-Create Date: 2014-06-20 09:01:52.704794
-
-"""
-
+from alembic import context
 from alembic.op import create_index, drop_index
 
-# revision identifiers, used by Alembic.
-revision = '35ef10d1e11b'  # pylint:disable=invalid-name
-down_revision = '3152492b110b'  # pylint:disable=invalid-name
+
+# Alembic revision identifiers
+revision = '35ef10d1e11b'
+down_revision = '3152492b110b'
 
 
 def upgrade():
     '''
-    upgrade method
+    Upgrade the database to this revision
     '''
-    create_index('REQUESTS_TYP_STA_UPD_IDX', 'requests', ["request_type", "state", "updated_at"])
-    drop_index('REQUESTS_TYP_STA_CRE_IDX', 'requests')
+
+    if context.get_context().dialect.name in ['oracle', 'mysql', 'postgresql']:
+        create_index('REQUESTS_TYP_STA_UPD_IDX', 'requests', ["request_type", "state", "updated_at"])
+        drop_index('REQUESTS_TYP_STA_CRE_IDX', 'requests')
 
 
 def downgrade():
     '''
-    downgrade method
+    Downgrade the database to the previous revision
     '''
-    create_index('REQUESTS_TYP_STA_CRE_IDX', 'requests', ["request_type", "state", "created_at"])
-    drop_index('REQUESTS_TYP_STA_UPD_IDX', 'requests')
+
+    if context.get_context().dialect.name in ['oracle', 'mysql', 'postgresql']:
+        create_index('REQUESTS_TYP_STA_CRE_IDX', 'requests', ["request_type", "state", "created_at"])
+        drop_index('REQUESTS_TYP_STA_UPD_IDX', 'requests')
